LayoutTests/imported/w3c/web-platform-tests/css/mediaqueries/navigation-controls.tentative-expected.txt


FAIL Should be known: '(navigation-controls)' assert_true: Can parse with JS expected true got false
FAIL Should be known: '(navigation-controls: none)' assert_true: Can parse with JS expected true got false
FAIL Should be known: '(navigation-controls: back-button)' assert_true: Can parse with JS expected true got false
FAIL Should be parseable: '(navigation-controls: none back-button)' assert_true: Can parse with JS expected true got false
PASS Should be unknown: '(navigation-controls: none back-button)'
FAIL Should be parseable: '(navigation-controls: back-button/none)' assert_true: Can parse with JS expected true got false
PASS Should be unknown: '(navigation-controls: back-button/none)'
FAIL Check that none evaluates to false in the boolean context assert_equals: expected true but got false
